Refactor plug.cli into a package
