+++ src/repobee_plug/cli/__init__.py
@@ -1,0 +1,21 @@
+from .settings import command_settings, command_extension_settings
+from .categorization import category
+from .args import option, positional, mutually_exclusive_group, ArgumentType
+from .commandmarkers import Command, CommandExtension
+
+from ._corecommand import _CoreCommand
+
+CoreCommand = _CoreCommand()
+
+__all__ = [
+    "ArgumentType",
+    "option",
+    "positional",
+    "mutually_exclusive_group",
+    "category",
+    "command_settings",
+    "command_extension_settings",
+    "Command",
+    "CommandExtension",
+    "CoreCommand",
+]
